service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.4
sentence-transformers==2.2.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
